requirements.txt
PyYAML>=6.0,<7.0
pytest>=8.0,<9.0
websockets>=13.0,<16.0
